Say who answers for a term wherever the three things are named
The reading was changed to answer a total out of the run it is taken over, and
four places still said a term's own reach is what its position was read to
hold. Two of them are the contract: the boundary's own account of what a part
comes to, and the ADR that decided a form is asked through one boundary. The
other two are the ADR index and the test that holds the three apart.

All four now say where the values it is answered from leave it, and the two
that carry the contract say who the two answerers are — a position's published
reach for a number one position answers, and what every value a run walks
guarantees for a total taken over one.

The ADR gains the reason as well, because it is the same decision's own defect
one axis along: several positions composing to one form is the space
direction, several occurrences accumulating to one term is the multiplicity
direction, and the second was open while the first was closed. And what the
run's answerer may use is what holds of every value it walks, never what a
context leaves them — a run fixes no case, so a reading narrowed by which case
a value turned out to be is a bound half the occurrences are outside.

The candidate generation's own tests move to the package it now lives in. They
were reaching a class in `check` that is no longer there, which a clean build
says and a local one with the class still sitting in `target` does not.
